release: v0.3.2
- cookbook: technical-deliverable channel in the tells library (1dca76c)
- approve honors CALLUS_CORPUS (#1)
- ruff RUF046 fix in the launch-GIF script
- version bumped in pyproject.toml, callus/__init__.py and the smoke test

+++ scripts/render_launch_gif.py
@@ -65,7 +65,7 @@ FONT_BADGE = _try_load_font(11, bold=True)
 
 def _lerp(a, b, t: float):
     t = max(0.0, min(1.0, t))
-    return tuple(int(round(a[i] + (b[i] - a[i]) * t)) for i in range(3))
+    return tuple(round(a[i] + (b[i] - a[i]) * t) for i in range(3))
 
 
 def _new_frame() -> Image.Image:
